requirements.txt
click==7.1.1
oauthlib===3.1.0
requests-oauthlib===1.3.0
requests-toolbelt===0.9.1
tqdm==4.45.0
colorama==0.4.3
watchdog==0.10.2